Add files via upload
A project based on sql database and jdbc connection. It consists of a login and register page which help in login of a member based on sql records and in register, a member can register their details which further register their details in sql database. Moreover there is also an option to update, delete or search a particular record.